I had Claude generate a lot of test data for benchmarking.
The next thing that needs to happen is making this somewhat efficient
when running for a long time with large files, mostly by allocating
a whole lot less than I'm doing now. It'd be fun to get some before
and after numbers to see how much of an effect it has.

Cow, time to learn your ways.