Massive refactoring of stdin handling.
Thanks, I hate it. But I'm handling stdin offsets for *small*
memory use situations now without too much of a mess. I have to
read some docs on what the limits are on line length for stdin so
I know what the limits I have to handle are. tmpfiles are in my
near future.

Made Clod summarize the current status.