chore: reworked config for easier testing
The test and non-test implementations of configuration now vary.
Test config is thread-local so you can change it per test without
worry, allowing you to set up whatever you want for a test without
interfering with other tests. And by "you", I mean "me".

Updated the cli help and the README to describe the current state
of the project.